Add AGPLv3 License
This introduces the long overdue license file into the repository and
adjusts the license field in the pyproject.toml. I choose AGPLv3 as it
retains most of our freedoms. Anyone hosting a (derived) instance of
this code should have a real incentive to contribute (or at least
publish) changes made to the source code. We are all doing this for the
FOSS community and not for some commerical entity that wants to sell
this as a service. That being said, even with the AGPLv3, there can be
commercial offerings that host this service. But they will have to share
their changes with their users which is good enough for me.