Reduce eval time by switching to gitignore.nix
The nix-gitignore implementation in Nixpkgs is very inefficient. It did
dominate our eval times. Almost all of the 23s (on my machine) to
instantiate the default.nix of this project could be attributed it.
After switching to the hercules-ci variant of the same (which is also
composable) the eval time is at around ~3s on my machine. Which is not
ideal but already a HUGE performance improvement. This is especially
important as with .envrc I am evaluating every time I'm dropping into
the shell.