Add a __str__ instance for GitHubEvent for better terminal formatting
In one of the previous commits we introduced the management command to
list matching events on the CLI but there we couldn't use any proper
formatting of the even and had to resort to just using the kind
attribute. This is a small step towards nicer formatting of an event.