tracker/models/GitHubEvent: show the GitHub PR/Issue id in __str__
Seeing that there is an even is already great but knowing *where* that
event related to is even better. With this commit we are showing the GH
Issue or PR number in the string representation of an GitHubEvent.

While at it also add the pk (id) of the event to the str representation.