Add import_channels systemd service & NixOS VM test
This is a very simple test case that just passes some channel to the
import command (via a local minio server) without actualy verifying the
data that is in the server once the command did succeed.

We can (and should) add verification once we have some public API to
query the packages that have been imported. For now we must trust that
the unittests are good enough in this regard.